chore(app-check): Refactor app check error logic
